Remove const. Create data.js for modules. Code review
