+++ src/view/filters/data.js
@@ -1,0 +1,5 @@
+const FILTERS = [`Everything`, `Future`, `Past`];
+
+export {
+  FILTERS,
+};
