+++ src/view/point-message/data.js
@@ -1,0 +1,8 @@
+const PointMessage = {
+  LOADING: `Loading...`,
+  NO_EVENTS: `Click New Event to create your first point`,
+};
+
+export {
+  PointMessage,
+};
