+++ src/view/sort/data.js
@@ -1,0 +1,5 @@
+const SORTS = [`Event`, `Time`, `Price`];
+
+export {
+  SORTS,
+};
