+++ src/view/tabs/data.js
@@ -1,0 +1,5 @@
+const TABS = [`Table`, `Stats`];
+
+export {
+  TABS,
+};
